requirements.txt
flask
flask-cors
google-cloud-firestore
firebase-admin
python-dotenv